# design/APISpec.yaml
openapi: 3.0.0
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/KASHYR26/training-webapp/1.0.0
info:
  version: "1.0.0"
  title: training-webapp
  description: The API for the training tracker webapp
paths:
  /users:
    get:
      tags:
        - Users
      description: Returns all registered users
      operationId: getAllUsers
      responses:
        '200':
          description: All the users
          content:
            application/json:
              schema:
                type: array
                items:
                    $ref: '#/components/schemas/ResponseUser'

        '403':
          $ref: '#/components/responses/403'

    post:
      tags:
        - Users
      operationId: registerUser
      description: Create new user
      responses:
        '201':
          description: Successfully registered user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseUser'
                
        '400':
          $ref: '#/components/responses/400'
                
        '409':
          $ref: '#/components/responses/409'
                
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReqBodyCreateUser'
  /users/{userId}:
    parameters:
      - name: userId
        in: path
        required: true
        description: User id
        schema:
          type: integer
    get:
      tags:
        - Users
      description: Returns the user specified by id
      operationId: getUserById
      responses:
        '200':
          description: Single user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseUser'
                
        '400':
          $ref: '#/components/responses/400'

        '403':
          $ref: '#/components/responses/403'

        '404':
          $ref: '#/components/responses/404'
    put:
      tags:
        - Users
      operationId: updateUser
      description: Update existing user
      responses:
        '200':
          description: Successfully updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseUser'
                
        '400':
          $ref: '#/components/responses/400'
                
        '401':
          $ref: '#/components/responses/401'

        '403':
          $ref: '#/components/responses/403'

        '404':
          $ref: '#/components/responses/404'
                
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReqBodyUpdateUser'
    delete:
      tags:
        - Users
      description: 'Delete the user specified by id'
      responses:
        '200':
          description: Deletes users and returns it
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseUser'

        '400':
          $ref: '#/components/responses/400'

        '401':
          $ref: '#/components/responses/401'

        '403':
          $ref: '#/components/responses/403'

        '404':
          $ref: '#/components/responses/404'
  /exercises:
    get:
      tags:
        - Exercises
      description: Returns all exercises
      operationId: getAllExercises
      responses:
        '200':
          description: List of all exercises
          content:
            application/json:
              schema:
                type: array
                items:
                    $ref: '#/components/schemas/ResponseExercise'
    post:
      tags:
        - Exercises
      description: Create new exercise
      operationId: createNewExercise
      responses:
        '201':
          description: Successfully created exercise
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseExercise'
                
        '400':
          $ref: '#/components/responses/400'
                
        '401':
          $ref: '#/components/responses/401'

        '403':
          $ref: '#/components/responses/403'
                
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReqBodyCreateExercise'
  /exercises/{exerciseId}:
    parameters:
      - name: exerciseId
        in: path
        required: true
        description: Exercise id
        schema:
          type: integer
    get:
      tags:
        - Exercises
      description: Returns the exercise specified by id
      operationId: getExerciseById
      responses:
        '200':
          description: Single exercise
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseExercise'
                
        '400':
          $ref: '#/components/responses/400'

        '404':
          $ref: '#/components/responses/404'

    put:
      tags:
        - Exercises
      description: Update exercise specified by id
      operationId: updateExercise
      responses:
        '200':
          description: Successfully updated exercise
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseExercise'
                
        '400':
          $ref: '#/components/responses/400'
                
        '401':
          $ref: '#/components/responses/401'

        '403':
          $ref: '#/components/responses/403'
                
        '404':
          $ref: '#/components/responses/404'
                
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReqBodyUpdateExercise'

    delete:
      tags:
        - Exercises
      description: 'Delete the exercise specified by id'
      responses:
        '200':
          description: Deletes exercise and returns it
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseExercise'

        '400':
          $ref: '#/components/responses/400'

        '401':
          $ref: '#/components/responses/401'

        '403':
          $ref: '#/components/responses/403'

        '404':
          $ref: '#/components/responses/404'
...
  /workouts/{workoutId}/exercises/{exerciseId}:
    parameters:
      - name: workoutId
        in: path
        required: true
        description: Workout id
        schema:
          type: integer
      - name: exerciseId
        in: path
        required: true
        description: Exercise id
        schema:
          type: integer

    put:
      tags:
        - Workouts
      description: Update exercise from workout
      operationId: updateExerciseFromWorkout
      responses:
        '200':
          description: Successfully updated exercise
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateExerciseFromWorkout'
              examples:
                Workout:
                  $ref: '#/components/examples/UpdateExerciseFromWorkout'
                
        '400':
          $ref: '#/components/responses/400'
                
        '401':
          $ref: '#/components/responses/401'

        '403':
          $ref: '#/components/responses/403'

        '404':
          $ref: '#/components/responses/404'
                
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReqBodyUpdateExerciseFromWorkout'
            examples:
              UpdateWorkoutExercise:
                $ref: '#/components/examples/ReqBodyUpdateExerciseFromWorkout'

    delete:
      tags:
        - Workouts
      description: Delete exercise with all of its sets from workout
      operationId: deleteWorkoutExercise
      responses:
        '200':
          description: Successfully deleted exercise from workout
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseAddExerciseToWorkout'
              examples:
                ExerciseWithAllItsSets:
                  type: array
                  $ref: '#/components/examples/ExerciseWithMultipleSets'
                
        '401':
          $ref: '#/components/responses/401'

        '403':
          $ref: '#/components/responses/403'

        '404':
          $ref: '#/components/responses/404'
  /workouts/{workoutId}/exercises/{exerciseId}/{exerciseSet}:
    parameters:
      - name: workoutId
        in: path
        required: true
        description: Workout id
        schema:
          type: integer
      - name: exerciseId
        in: path
        required: true
        description: Exercise id
        schema:
          type: integer
      - name: exerciseSet
        in: path
        required: true
        description: Exercise set
        schema:
          type: integer

    delete:
      tags:
        - Workouts
      description: Delete exercise set from workout
      operationId: deleteWorkoutExerciseSet
      responses:
        '200':
          description: Successfully deleted exercise set from workout
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseAddExerciseToWorkout'
              examples:
                Workout:
                  $ref: '#/components/examples/ExerciseFromWorkout'
                
        '401':
          $ref: '#/components/responses/401'

        '403':
          $ref: '#/components/responses/403'

        '404':
          $ref: '#/components/responses/404'
...
  /workouts/templates/{templateId}/exercises/{exerciseId}:
    parameters:
      - name: templateId
        in: path
        required: true
        description: Template id
        schema:
          type: integer
      - name: exerciseId
        in: path
        required: true
        description: Exercise id
        schema:
          type: integer

    put:
      tags:
        - Workouts
      description: Updates exercise from workout template
      operationId: updateExerciseToWorkoutTemplate
      responses:
        '200':
          description: Successfully updated exercise to template
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseWorkoutTemplateExercise'
              examples:
                UpdatedExercise:
                  $ref: '#/components/examples/WorkoutTemplateExercise'
                
        '400':
          $ref: '#/components/responses/400'
                
        '401':
          $ref: '#/components/responses/401'

        '403':
          $ref: '#/components/responses/403'

        '404':
          $ref: '#/components/responses/404'
                
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReqBodyUpdateExerciseFromWorkoutTemplate'
            examples:
              UpdateExerciseToWorkoutTemplate:
                $ref: '#/components/examples/UpdateExerciseToWorkoutTemplate'
  /workouts/templates/{templateId}/exercises/{exerciseId}/{exerciseOrder}:
    parameters:
      - name: templateId
        in: path
        required: true
        description: Template id
        schema:
          type: integer
      - name: exerciseId
        in: path
        required: true
        description: Exercise id
        schema:
          type: integer
      - name: exerciseOrder
        in: path
        required: true
        description: Exercise order
        schema:
          type: integer

    delete:
      tags:
        - Workouts
      description: Deletes exercise from workout template
      operationId: deleteExerciseToWorkoutTemplate
      responses:
        '200':
          description: Successfully deleted exercise to template
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseWorkoutTemplateExercise'
              examples:
                UpdatedExercise:
                  $ref: '#/components/examples/WorkoutTemplateExercise'
                
        '400':
          $ref: '#/components/responses/400'
                
        '401':
          $ref: '#/components/responses/401'

        '403':
          $ref: '#/components/responses/403'

        '404':
          $ref: '#/components/responses/404'
components:
  schemas:
    ResponseUser:
      type: object
      properties:
        id:
          type: integer
          description: User id
        alias:
          type: string
          description: User name
        email:
          type: string
          description: User email
        last_name:
          type: string
          description: User last name
        img:
          type: string
          description: User profile image's URI
        second_last_name:
          type: string
          description: User second last name
      example:
        id: 3
        alias: "John"
        email: 'John.Doe@domain.com'
        last_name: 'Doe'
        img: 'https://i.pinimg.com/736x/7f/64/3f/7f643f0db514d7971349c416e29e42a8.jpg'
        second_last_name: 'Smith'
      description: User object
    ReqBodyCreateUser:
      type: object
      required:
        - alias
        - email
        - password
      properties:
        alias:
          type: string
          description: User name
        email:
          type: string
          description: User email
        password:
          type: string
          description: User password
        last_name:
          type: string
          description: User last name
        second_last_name:
          type: string
          description: User second last name
      example:
        alias: "John"
        email: 'John.Doe@domain.com'
        last_name: 'Doe'
        password: 'secure_password'
        second_last_name: 'Smith'
      description: Info needed to create user object
    ReqBodyUpdateUser:
      type: object
      properties:
        alias:
          type: string
          description: User name
        email:
          type: string
          description: User email
        password:
          type: string
          description: User password
        last_name:
          type: string
          description: User last name
        second_last_name:
          type: string
          description: User second last name
        img:
          type: string
          description: User profile image

      example:
        alias: "John"
        email: 'John.Doe@domain.com'
        last_name: 'Doe'
        password: 'secure_password'
        second_last_name: 'Smith'
        # TODO research about how to handle it from server. This is susceptible of beign changed
        img: "Link to img"
      description: Info needed to update user object
    ResponseExercise:
      type: object
      properties:
        id:
          type: integer
          description: Exercise id
        alias:
          type: string
          description: Exercise name
        description:
          type: string
          description: Exercise email
      example:
        id: 3
        alias: "Bicep curl"
        description: "A staple in strength training, this exercise targets the biceps brachii muscle group. Executed with dumbbells or a barbell, the movement involves flexing the elbow to lift the weight towards the shoulders while keeping the upper arms stationary. Proper form is crucial to isolate the biceps and prevent swinging or momentum. It strengthens the biceps, improves arm aesthetics, and aids in functional movements like lifting."
      description: Exercise object
    ResponseWorkoutTemplateExercise:
      type: object
      properties:
        workoutTemplateId:
          type: integer
          description: Workout template's id that owns the exercise
        exerciseId:
          type: integer
          description: Exercise id
        exerciseOrder:
          type: integer
          description: Order in which the exercise is performed
        exerciseSets:
          type: integer
          description: Number of sets of the exercise to perform
      example:
        id: 3
        alias: "Bicep curl"
        description: "A staple in strength training, this exercise targets the biceps brachii muscle group. Executed with dumbbells or a barbell, the movement involves flexing the elbow to lift the weight towards the shoulders while keeping the upper arms stationary. Proper form is crucial to isolate the biceps and prevent swinging or momentum. It strengthens the biceps, improves arm aesthetics, and aids in functional movements like lifting."
      description: Exercise object
    ReqBodyCreateExercise:
      type: object
      required:
        - alias
      properties:
        alias:
          type: string
          description: Exercise name
        description:
          type: string
          description: Exercise email
      example:
        alias: "Bicep curl"
        description: "A staple in strength training, this exercise targets the biceps brachii muscle group. Executed with dumbbells or a barbell, the movement involves flexing the elbow to lift the weight towards the shoulders while keeping the upper arms stationary. Proper form is crucial to isolate the biceps and prevent swinging or momentum. It strengthens the biceps, improves arm aesthetics, and aids in functional movements like lifting."
      description: Exercise object
    ReqBodyUpdateExercise:
      type: object
      properties:
        alias:
          type: string
          description: Exercise name
        description:
          type: string
          description: Exercise email
      example:
        alias: "Bicep curl"
        description: "A staple in strength training, this exercise targets the biceps brachii muscle group. Executed with dumbbells or a barbell, the movement involves flexing the elbow to lift the weight towards the shoulders while keeping the upper arms stationary. Proper form is crucial to isolate the biceps and prevent swinging or momentum. It strengthens the biceps, improves arm aesthetics, and aids in functional movements like lifting."
      description: Exercise object
...
    ReqBodyUpdateExerciseFromWorkout:
      type: object
      required:
        - exerciseSet
      properties:
        exerciseSet:
          type: integer
          description: Number of the set in which the exercise was performed
        reps:
          type: integer
          description: Number of reps performed
        weight:
          type: number
          description: Weight used
        time_in_seconds:
          type: integer
          description: Time in seconds that the exercise spanned
    ResponseAddExerciseToWorkout:
      type: object
      required:
        - exerciseId
        - exerciseSet
      properties:
        exerciseId:
          type: integer
          description: Id of the exercise
        exerciseSet:
          type: integer
          description: Number of the set in which the exercise was performed
        reps:
          type: integer
          description: Number of reps performed
        weight:
          type: number
          description: Weight used
        time_in_seconds:
          type: integer
          description: Time in seconds that the exercise spanned
    UpdateExerciseFromWorkout:
      type: object
      required:
        - exerciseSet
      properties:
        exerciseSet:
          type: integer
          description: Number of the set in which the exercise was performed
        reps:
          type: integer
          description: Number of reps performed
        weight:
          type: number
          description: Weight used
        time_in_seconds:
          type: integer
          description: Time in seconds that the exercise spanned
...
    ReqBodyUpdateExerciseFromWorkoutTemplate:
      type: object
      properties:
        exerciseOrder:
          type: integer
          description: Order of the exercise in the template
        exerciseSets:
          type: integer
          description: Number of sets of the exercise to perform
...
  responses:
    400:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              msg:
                type: string
                description: Error messagge. 
            example:
              msg: Wrong or missing parameter
    401:
      description: Not logged in
      content:
        application/json:
          schema:
            type: object
            properties:
              msg:
                type: string
                description: Error messagge. 

            example:
              msg: Need to log in
    403:
      description: Not having permissions to access information
      content:
        application/json:
          schema:
            type: object
            properties:
              msg:
                type: string
                description: Error messagge. 
          example:
            msg: Don't have permissions to access information
    404:
      description: Resource was not found
      content:
        application/json:
          schema:
            type: object
            properties:
              msg:
                type: string
                description: Error messagge. 

            example:
              msg: Resource was not found
    409:
      description: Already existing
      content:
        application/json:
          schema:
            type: object
            properties:
              msg:
                type: string
                description: Error messagge. 

            example:
              msg: Already existing resource
  examples:
...
    WorkoutTemplateExercise:
      value:
        workoutTemplateId: 2
        exerciseId: 7
        exerciseOrder: 2
        exerciseSets: 3
...
    ExerciseFromWorkout:
      value:
        exerciseId: 1
        exerciseSet: 2
        reps: 7
        weight: 80
        time_in_seconds: null
    ExerciseWithMultipleSets:
      value:
        - exerciseId: 1
          exerciseSet: 1
          reps: 7
          weight: 80
          time_in_seconds: null
        - exerciseId: 1
          exerciseSet: 2
          reps: 7
          weight: 80
          time_in_seconds: null
        - exerciseId: 1
          exerciseSet: 3
          reps: 6
          weight: 80
          time_in_seconds: null
    ReqBodyUpdateExerciseFromWorkout:
      value:
        exerciseSet: 2
        reps: 7
        weight: 80
        time_in_seconds: null
    UpdateExerciseFromWorkout:
      value:
        exerciseId: 1
        exerciseSet: 2
        reps: 7
        weight: 80
        time_in_seconds: null
...
    UpdateExerciseToWorkoutTemplate:
      value:
        exerciseOrder: 6
        exerciseSets: 4